shared pointer support for Scope in Expression
this is not very useful at the moment, because we don't have shared_ptr
elsewhere in the code. It is simply a first step towards fixing the
memory issues in the code base.

A good next step is to start introducing shared_ptr into all
expressions, and start working through the unit tests from simple to
more complicated expressions.